+++ breeze-serialization/src/main/kotlin/extension/SerializationInternalExtensions.kt
@@ -9,10 +9,11 @@ package com.windea.breezeframework.serialization.extension
 import com.windea.breezeframework.core.extension.*
 import com.windea.breezeframework.serialization.*
 import com.windea.breezeframework.serialization.component.*
+import java.io.*
 
-internal inline fun Char.writeTo(writer: DataWriter) = writer.write(this)
+internal inline fun Char.appendTo(buffer: Appendable) = buffer.append(this)
 
-internal inline fun String.writeTo(writer: DataWriter) = writer.write(this)
+internal inline fun String.appendTo(buffer: Appendable) = buffer.append(this)
 
 
 internal val defaultMapLikeSerializer: MapLikeSerializer = MapLikeSerializer.BreezeMapLikeSerializer()